Add traits to update/remove components
I plan to extend the macro to also generate the handle enums, but those
handle enums will have to work with multiple generated components
collections, not just the ones that were generated alongside them. For
example, the component enum used for shared components will have to be
able to remove components from the client components collection.

The infrastructure added in this commit will enable this
inter-operation.