Enable bespoke frontend unconditionally
I'm not sure if this is going to work out, but the problems we had
before were related to cross-compiling to the x86_64-pc-windows-gnu
target. We're no longer cross-compiling, and we're no longer using that
target, so it's possible this just works now.

If it doesn't, I'll have to revert this commit.